+++ magda-authorization-api/src/Database.ts
@@ -274,7 +274,9 @@ export default class Database {
             user.managingOrgUnitIds = [];
             user.orgUnit = null;
         } else {
-            user.orgUnit = await this.orgQueryer.getNodeById(user.orgUnitId);
+            user.orgUnit = (await this.orgQueryer.getNodeById(
+                user.orgUnitId
+            )).valueOr(null);
             user.managingOrgUnitIds = (await this.orgQueryer.getAllChildren(
                 user.orgUnitId,
                 true,